minor: move tests.contents to tests.testing_utils.mocks.
